requirements.txt
requests
daphne
channels
channels-redis
Django
pymysql
mysqlclient
django-simple-captcha
django-import-export
django-simpleui
django_redis
librosa
numpy
opencv-python
paddlepaddle
paddlenlp
ppasr
sshtunnel
tensorflow_cpu==2.9.3